+++ src/compas_timber/ghpython/components_cpython/CT_Joint_Rule_Topology_T/code.py
@@ -29,18 +29,13 @@ class T_TopologyJointRule(Grasshopper.Kernel.GH_ScriptInstance):
             ghenv.Component.Message = "Default: TButtJoint"
             ghenv.Component.AddRuntimeMessage(Grasshopper.Kernel.GH_RuntimeMessageLevel.Warning, "TButtJoint is default, change in context menu (right click)")
             return TopologyRule(JointTopology.TOPO_T, TButtJoint)
-        else:
-            ghenv.Component.Message = self.joint_type.__name__
-            kwargs = {}
-            for i, val in enumerate(args):
-                if val is not None:
-                    kwargs[self.arg_names()[i]] = val
-            supported_topo = self.joint_type.SUPPORTED_TOPOLOGY
-            if not isinstance(supported_topo, list):
-                supported_topo = [supported_topo]
-            if JointTopology.TOPO_T not in supported_topo:
-                ghenv.Component.AddRuntimeMessage(Grasshopper.Kernel.GH_RuntimeMessageLevel.Warning, "Joint type does not match topology. Joint may not be generated.")
-            return TopologyRule(JointTopology.TOPO_T, self.joint_type, **kwargs)
+
+        ghenv.Component.Message = self.joint_type.__name__
+        kwargs = {}
+        for i, val in enumerate(args):
+            if val is not None:
+                kwargs[self.arg_names()[i]] = val
+        return TopologyRule(JointTopology.TOPO_T, self.joint_type, **kwargs)
 
     def arg_names(self):
         names = inspect.getargspec(self.joint_type.__init__)[0][3:]